+++ cai_causal_graph/causal_graph.py
@@ -447,6 +447,8 @@ class CausalGraph(HasIdentifier, HasMetadata, CanDictSerialize, CanDictDeseriali
             inputs connected to all outputs. If no `input_list` and no `output_list` is provided, an empty graph will
             be created. If either or both are provided, but this is `False`, then the nodes will be added but not
             connected by edges.
+        :param meta: Any metadata defined on the graph. The keys must be strings, but no requirement is placed on the
+            values of the dictionary.
         """
         self._nodes_by_identifier: Dict[str, Node] = dict()
         self._edges_by_source: Dict[str, Dict[str, Edge]] = defaultdict(dict)
