+++ cai_causal_graph/interfaces.py
@@ -16,9 +16,13 @@ limitations under the License.
 from __future__ import annotations
 
 import abc
-from typing import Optional
+import logging
+from copy import copy
+from typing import List, Optional
 
-from cai_causal_graph.metadata_handler import HasMeta
+from cai_causal_graph.metadata_handler import HasMeta, MetaDataError, MetaField
+
+logger = logging.getLogger(__name__)
 
 
 class CanDictSerialize(abc.ABC):
@@ -56,6 +60,75 @@ class HasIdentifier(abc.ABC):
 
 class HasMetadata(HasMeta):
     """An interface for an object that has metadata."""
+
+    def __init__(self, *args, meta: Optional[dict] = None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.meta = meta.copy() if meta is not None else dict()
+        assert isinstance(self.meta, dict) and all(
+            isinstance(k, str) for k in self.meta
+        ), 'Metadata must be provided as a dictionary with strings as keys.'
+
+    @classmethod
+    def get_metadata_schema(cls) -> List[MetaField]:
+        return []
+
+    @classmethod
+    def _get_field_for_property_name(cls, property_name: str):
+        schema = cls.get_metadata_schema()
+        cls._validate_schema(schema=schema)
+
+        matched_fields = list(filter(lambda field_: field_.property_name == property_name, schema))
+
+        if len(matched_fields) == 0:
+            raise KeyError(f'No metadata field with property name matching {property_name} in the schema {schema}.')
+        elif len(matched_fields) > 1:
+            # This should never be raised in reality since schema is validated.
+            raise MetaDataError(f'Found multiple fields {matched_fields} with identical property name.')
+        else:
+            return matched_fields[0]
+
+    @classmethod
+    def _process_meta(
+        cls, meta: Optional[dict], kwargs_dict: dict, raise_if_unknown_tags: bool = False
+    ) -> Optional[dict]:
+        schema = cls.get_metadata_schema()
+        cls._validate_schema(schema=schema)
+
+        requested_meta = {field.metatag: kwargs_dict.pop(field.parameter_name, field.default_value) for field in schema}
+
+        if raise_if_unknown_tags and len(kwargs_dict) > 0:
+            raise MetaDataError(f'Unknown keyword arguments {kwargs_dict}. Metadata schema is {schema}.')
+
+        return cls._update_metadata(meta=copy(meta), **requested_meta)
+
+    @classmethod
+    def _validate_schema(cls, schema: List[MetaField]):
+        tags, properties, parameters = list(
+            zip(*[[field.metatag, field.property_name, field.parameter_name] for field in schema])
+        )
+
+        for l, name in [(tags, 'metatags'), (properties, 'property names'), (parameters, 'parameter names')]:
+            if len(set(l)) != len(l):
+                raise MetaDataError(f'Found multiple meta fields with identical {name} in schema: {schema}.')
+
+    @classmethod
+    def _update_metadata(cls, meta: Optional[dict], **kwargs) -> Optional[dict]:
+        if meta is None:
+            meta = dict()
+
+        for k, v in kwargs.items():
+            if v is not None:
+                if (existing_v := meta.get(k, None)) is not None:
+                    logger.debug(
+                        f'{k} (set to {existing_v}) in the metadata of {cls.__name__} is '
+                        f'overwritten by the newly provided value {v}.'
+                    )
+                meta[k] = v
+
+        if len(meta) == 0:
+            return None
+
+        return meta
 
     @property
     def metadata(self) -> Optional[dict]:
